requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
fire==0.3.1
gql==3.0.0a5
graphql-core==3.1.2
idna==2.10
multidict==5.1.0
mypy-extensions==0.4.3
packageurl-python==0.9.3
pathspec==0.8.1
promise==2.3
pybraries==0.3.4
python-dateutil==2.8.1
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
Rx==1.6.1
six==1.15.0
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
websockets==8.1
yarl==1.6.3
